+++ src/NewPicker/PickerInput/Selector/Input.tsx
@@ -5,7 +5,7 @@ import raf from 'rc-util/lib/raf';
 import * as React from 'react';
 import { leftPad } from '../../../utils/miscUtil';
 import PickerContext from '../../../PickerInput/context';
-import useLockEffect from '../hooks/useLockEffect';
+import useLockEffect from '../../../PickerInput/hooks/useLockEffect';
 import Icon from './Icon';
 import MaskFormat from './MaskFormat';
 import { getMaskRange } from './util';
