+++ src/PickerInput/SinglePicker.tsx
@@ -17,16 +17,16 @@ import type {
 } from '../interface';
 import PickerTrigger from '../PickerTrigger';
 import PickerContext from './context';
-import useCellRender from '../NewPicker/PickerInput/hooks/useCellRender';
-import useFieldsInvalidate from '../NewPicker/PickerInput/hooks/useFieldsInvalidate';
-import useFilledProps from '../NewPicker/PickerInput/hooks/useFilledProps';
-import useOpen from '../NewPicker/PickerInput/hooks/useOpen';
-import { usePickerRef } from '../NewPicker/PickerInput/hooks/usePickerRef';
-import usePresets from '../NewPicker/PickerInput/hooks/usePresets';
-import useRangeActive from '../NewPicker/PickerInput/hooks/useRangeActive';
-import useRangePickerValue from '../NewPicker/PickerInput/hooks/useRangePickerValue';
-import useRangeValue, { useInnerValue } from '../NewPicker/PickerInput/hooks/useRangeValue';
-import useShowNow from '../NewPicker/PickerInput/hooks/useShowNow';
+import useCellRender from './hooks/useCellRender';
+import useFieldsInvalidate from './hooks/useFieldsInvalidate';
+import useFilledProps from './hooks/useFilledProps';
+import useOpen from './hooks/useOpen';
+import { usePickerRef } from './hooks/usePickerRef';
+import usePresets from './hooks/usePresets';
+import useRangeActive from './hooks/useRangeActive';
+import useRangePickerValue from './hooks/useRangePickerValue';
+import useRangeValue, { useInnerValue } from './hooks/useRangeValue';
+import useShowNow from './hooks/useShowNow';
 import Popup from '../NewPicker/PickerInput/Popup';
 import SingleSelector from '../NewPicker/PickerInput/Selector/SingleSelector';
