+++ src/PickerInput/hooks/useFilledProps.ts
@@ -1,11 +1,11 @@
 import { warning } from 'rc-util';
 import * as React from 'react';
-import { toArray } from '../../../utils/miscUtil';
-import { fillLocale } from '../../../hooks/useLocale';
-import { getTimeConfig } from '../../../hooks/useTimeConfig';
-import type { FormatType, InternalMode } from '../../../interface';
-import type { RangePickerProps } from '../../../PickerInput/RangePicker';
-import { fillClearIcon } from '../Selector/hooks/useClearIcon';
+import { toArray } from '../../utils/miscUtil';
+import { fillLocale } from '../../hooks/useLocale';
+import { getTimeConfig } from '../../hooks/useTimeConfig';
+import type { FormatType, InternalMode } from '../../interface';
+import type { RangePickerProps } from '../RangePicker';
+import { fillClearIcon } from '../../NewPicker/PickerInput/Selector/hooks/useClearIcon';
 import useDisabledBoundary from './useDisabledBoundary';
 import { useFieldFormat } from './useFieldFormat';
 import useInputReadOnly from './useInputReadOnly';
