+++ src/PickerInput/hooks/useInvalidate.ts
@@ -1,11 +1,11 @@
 import { useEvent } from 'rc-util';
-import type { GenerateConfig } from '../../../generate';
+import type { GenerateConfig } from '../../generate';
 import type {
   PanelMode,
   RangeTimeProps,
   SharedPickerProps,
   SharedTimeProps,
-} from '../../../interface';
+} from '../../interface';
 
 /**
  * Check if provided date is valid for the `disabledDate` & `showTime.disabledTime`.
