+++ src/PickerInput/hooks/useRangePickerValue.ts
@@ -1,10 +1,10 @@
 import { useMergedState } from 'rc-util';
 import useLayoutEffect from 'rc-util/lib/hooks/useLayoutEffect';
 import * as React from 'react';
-import type { GenerateConfig } from '../../../generate';
-import { isSame } from '../../../utils/dateUtil';
-import type { InternalMode, Locale, PanelMode } from '../../../interface';
-import type { RangePickerProps } from '../../../PickerInput/RangePicker';
+import type { GenerateConfig } from '../../generate';
+import { isSame } from '../../utils/dateUtil';
+import type { InternalMode, Locale, PanelMode } from '../../interface';
+import type { RangePickerProps } from '../RangePicker';
 
 export function offsetPanelDate<DateType = any>(
   generateConfig: GenerateConfig<DateType>,
