+++ src/PickerInput/hooks/useRangeValue.ts
@@ -1,12 +1,12 @@
 import { useEvent, useMergedState } from 'rc-util';
 import * as React from 'react';
-import type { GenerateConfig } from '../../../generate';
-import useSyncState from '../../../hooks/useSyncState';
-import type { BaseInfo, FormatType, Locale, ReplaceListType } from '../../../interface';
-import { formatValue, isSame, isSameTimestamp } from '../../../utils/dateUtil';
-import { fillIndex } from '../../../utils/miscUtil';
-import type { RangePickerProps } from '../../../PickerInput/RangePicker';
-import type { ReplacedPickerProps } from '../../../PickerInput/SinglePicker';
+import type { GenerateConfig } from '../../generate';
+import useSyncState from '../../hooks/useSyncState';
+import type { BaseInfo, FormatType, Locale, ReplaceListType } from '../../interface';
+import { formatValue, isSame, isSameTimestamp } from '../../utils/dateUtil';
+import { fillIndex } from '../../utils/miscUtil';
+import type { RangePickerProps } from '../RangePicker';
+import type { ReplacedPickerProps } from '../SinglePicker';
 import useLockEffect from './useLockEffect';
 
 const EMPTY_VALUE: any[] = [];
